Add failsafe characters.
Prevent raising an error if some of the characters are
missing from the output encoding

If there is no character in the output encoding that
represent that in the input then place other charactars
that are visualy like the source character.

If there is no availabale failsafe character then replace it
by "?".